+++ packages/moonstone/ContextualPopupDecorator/ContextualPopup.js
@@ -66,13 +66,13 @@ const ContextualPopupBase = kind({
 		/**
 		 * The element(s) to be displayed in the body of the popup.
 		 *
-		 * @type {Node}
+		 * @type {Element|Element[]}
 		 * @required
 		 * @public
 		 */
 		children: PropTypes.oneOfType([
-			PropTypes.arrayOf(PropTypes.element),
-			PropTypes.element
+			PropTypes.element,
+			PropTypes.arrayOf(PropTypes.element)
 		]).isRequired,
 
 		/**
@@ -82,10 +82,10 @@ const ContextualPopupBase = kind({
 		 * @public
 		 */
 		arrowPosition: PropTypes.shape({
-			top: PropTypes.number,
 			bottom: PropTypes.number,
 			left: PropTypes.number,
-			right: PropTypes.number
+			right: PropTypes.number,
+			top: PropTypes.number
 		}),
 
 		/**
@@ -95,10 +95,10 @@ const ContextualPopupBase = kind({
 		 * @public
 		 */
 		containerPosition: PropTypes.shape({
-			top: PropTypes.number,
 			bottom: PropTypes.number,
 			left: PropTypes.number,
-			right: PropTypes.number
+			right: PropTypes.number,
+			top: PropTypes.number
 		}),
 
 		/**