+++ packages/moonstone/Divider/Divider.js
@@ -1,29 +1,31 @@
 /**
- * Exports the {@link moonstone/Divider.Divider} component.
+ * Provides Moonstone-themed divider components and behaviors.
  *
  * @module moonstone/Divider
+ * @exports Divider
+ * @exports DividerBase
+ * @exports DividerDecorator
  */
 
 import kind from '@enact/core/kind';
-import Uppercase from '@enact/i18n/Uppercase';
-import React from 'react';
 import PropTypes from 'prop-types';
 import Pure from '@enact/ui/internal/Pure';
+import Uppercase from '@enact/i18n/Uppercase';
+import compose from 'ramda/src/compose';
+import defaultProps from 'recompose/defaultProps';
+import setPropTypes from 'recompose/setPropTypes';
+import React from 'react';
 
 import {MarqueeDecorator} from '../Marquee';
 import Skinnable from '../Skinnable';
 
 import css from './Divider.less';
 
-const MarqueeH3 = Uppercase(MarqueeDecorator('h3'));
-
 /**
- * {@link moonstone/Divider.Divider} is a simply styled component that may be used as a separator
- * between groups of components.
+ * A simply styled component that may be used as a separator between groups of components.
  *
  * @class Divider
  * @memberof moonstone/Divider
- * @mixes moonstone/MarqueeDecorator.MarqueeDecorator
  * @ui
  * @public
  */
@@ -32,33 +34,15 @@ const DividerBase = kind({
 
 	propTypes: /** @lends moonstone/Divider.Divider.prototype */ {
 		/**
-		 * Configures how the `children` string will be capitalized. By default, each word is capitalized.
+		 * The divider content.
 		 *
-		 * @see i18n/Uppercase#casing
-		 * @type {String}
-		 * @default 'word'
+		 * A divider with no children (text content) will render simply as a horizontal line, with
+		 * even spacing above and below.
+		 *
+		 * @type {Node}
 		 * @public
 		 */
-		casing: PropTypes.oneOf(['upper', 'preserve', 'word', 'sentence']),
-
-		/**
-		 * The content of the divider. A divider with no children (text content) will render simply
-		 * as a horizontal line, with even spacing above and below.
-		 *
-		 * @type {String}
-		 * @public
-		 */
-		children: PropTypes.string,
-
-		/**
-		 * The children string will have each word capitalized, unless this is set to `true`.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @deprecated use `casing`
-		 * @public
-		 */
-		preserveCase: PropTypes.bool,
+		children: PropTypes.node,
 
 		/**
 		 * The size of the spacing around the divider.
@@ -71,7 +55,7 @@ const DividerBase = kind({
 		 *
 		 * _Note:_ Spacing is separate from margin with regard to `margin-top`. It ensures a
 		 * consistent distance from the bottom horizontal line. It's safe to use `margin-top` to add
-		 * additional spacing above your {@link moonstone/Divider.Divider}.
+		 * additional spacing above your `Divider`.
 		 *
 		 * @type {String}
 		 * @default 'normal'
@@ -81,8 +65,6 @@ const DividerBase = kind({
 	},
 
 	defaultProps: {
-		casing: 'word',
-		preserveCase: false,
 		spacing: 'normal'
 	},
 
@@ -95,21 +77,78 @@ const DividerBase = kind({
 		className: ({spacing, styler}) => styler.append(spacing)
 	},
 
-	render: ({children, ...rest}) => {
-		delete rest.spacing;
+	render: (props) => {
+		delete props.spacing;
 
 		return (
-			// TODO: change to `marqueeOn="render"`
-			<MarqueeH3 marqueeOn="hover" {...rest}>{children}</MarqueeH3>
+			<h3 {...props} />
 		);
 	}
 });
 
-const Divider = Pure(
-	Skinnable(
-		DividerBase
-	)
+/**
+ * Moonstone-specific divider behaviors to apply to [Divider]{@link moonstone/Divider.DividerBase}.
+ *
+ * @hoc
+ * @memberof moonstone/Divider
+ * @mixes i18n/Uppercase.Uppercase
+ * @mixes moonstone/MarqueeDecorator.MarqueeDecorator
+ * @mixes ui/Skinnable.Skinnable
+ */
+const DividerDecorator = compose(
+	setPropTypes({
+		marqueeOn: PropTypes.oneOf(['hover', 'render'])
+	}),
+	defaultProps({
+		casing: 'word',
+		marqueeOn: 'render'
+	}),
+	Pure,
+	Uppercase,
+	MarqueeDecorator,
+	Skinnable
 );
 
+/**
+ * A Moonstone-styled divider with built-in support for uppercasing, and marqueed text.
+ *
+ * @class Divider
+ * @extends moonstone/Divider.DividerBase
+ * @memberof moonstone/Divider
+ * @mixes moonstone/Divider.DividerDecorator
+ * @ui
+ * @public
+ */
+
+/**
+ * Capitalization style for `children`.
+ *
+ * @name casing
+ * @type {String}
+ * @default 'word'
+ * @memberof moonstone/Divider.Divider.prototype
+ * @see i18n/Uppercase#casing
+ * @public
+ */
+
+/**
+ * Marquee animation trigger.
+ *
+ * Valid values are `'hover'` and `'render'`.
+ *
+ * @name marqueeOn
+ * @type {String}
+ * @default 'render'
+ * @memberof moonstone/Divider.Divider.prototype
+ * @see moonstone/Marquee.Marquee
+ * @public
+ */
+
+const Divider = DividerDecorator(DividerBase);
+
 export default Divider;
-export {Divider, DividerBase};
+export {
+	Divider,
+	DividerBase,
+	DividerDecorator
+};