+++ packages/moonstone/Scrollable/Scrollable.js
@@ -7,10 +7,15 @@
  * @private
  */
 
+import classNames from 'classnames';
 import {constants, ScrollableBase as UiScrollableBase} from '@enact/ui/Scrollable';
-import {getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
+import {getDirection} from '@enact/spotlight';
+import {getTargetByDirectionFromElement, getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
+import {Job} from '@enact/core/util';
+import platform from '@enact/core/platform';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
+import ReactDOM from 'react-dom';
 import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 import Touchable from '@enact/ui/Touchable';
@@ -18,7 +23,9 @@ import Touchable from '@enact/ui/Touchable';
 import $L from '../internal/$L';
 
 import Scrollbar from './Scrollbar';
+import Skinnable from '../Skinnable';
 
+import overscrollCss from './OverscrollEffect.less';
 import scrollbarCss from './Scrollbar.less';
 
 const
@@ -26,8 +33,11 @@ const
 		animationDuration,
 		isPageDown,
 		isPageUp,
+		overscrollTypes,
 		paginationPageMultiplier
 	} = constants,
+	overscrollRatioPrefix = '--scrollable-overscroll-ratio-',
+	overscrollTimeout = 300,
 	reverseDirections = {
 		down: 'up',
 		left: 'right',
@@ -138,39 +148,39 @@ class ScrollableBase extends Component {
 		focusableScrollbar: PropTypes.bool,
 
 		/**
-		* Sets the hint string read when focusing the next button in the vertical scroll bar.
-		*
-		* @type {String}
-		* @default $L('scroll down')
-		* @public
-		*/
+		 * Sets the hint string read when focusing the next button in the vertical scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll down')
+		 * @public
+		 */
 		scrollDownAriaLabel: PropTypes.string,
 
 		/**
-		* Sets the hint string read when focusing the previous button in the horizontal scroll bar.
-		*
-		* @type {String}
-		* @default $L('scroll left')
-		* @public
-		*/
+		 * Sets the hint string read when focusing the previous button in the horizontal scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll left')
+		 * @public
+		 */
 		scrollLeftAriaLabel: PropTypes.string,
 
 		/**
-		* Sets the hint string read when focusing the next button in the horizontal scroll bar.
-		*
-		* @type {String}
-		* @default $L('scroll right')
-		* @public
-		*/
+		 * Sets the hint string read when focusing the next button in the horizontal scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll right')
+		 * @public
+		 */
 		scrollRightAriaLabel: PropTypes.string,
 
 		/**
-		* Sets the hint string read when focusing the previous button in the vertical scroll bar.
-		*
-		* @type {String}
-		* @default $L('scroll up')
-		* @public
-		*/
+		 * Sets the hint string read when focusing the previous button in the vertical scroll bar.
+		 *
+		 * @type {String}
+		 * @default $L('scroll up')
+		 * @public
+		 */
 		scrollUpAriaLabel: PropTypes.string
 	}
 
@@ -201,6 +211,13 @@ class ScrollableBase extends Component {
 		}
 	}
 
+	componentWillUnmount () {
+		this.stopOverscrollJob('horizontal', 'before');
+		this.stopOverscrollJob('horizontal', 'after');
+		this.stopOverscrollJob('vertical', 'before');
+		this.stopOverscrollJob('vertical', 'after');
+	}
+
 	// status
 	isWheeling = false
 
@@ -209,6 +226,17 @@ class ScrollableBase extends Component {
 	lastScrollPositionOnFocus = null
 	indexToFocus = null
 	nodeToFocus = null
+
+	// voice control
+	isVoiceControl = false
+	voiceControlDirection = 'vertical'
+
+	// overscroll
+	overscrollRefs = {horizontal: null, vertical: null}
+	overscrollJobs = {
+		horizontal: {before: null, after: null},
+		vertical: {before: null, after: null}
+	}
 
 	onFlick = () => {
 		const focusedItem = Spotlight.getCurrent();
@@ -240,15 +268,26 @@ class ScrollableBase extends Component {
 		if (pos) {
 			const
 				{top, left} = pos,
-				bounds = this.uiRef.getScrollBounds();
+				bounds = this.uiRef.getScrollBounds(),
+				scrollHorizontally = bounds.maxLeft > 0 && left !== this.uiRef.scrollLeft,
+				scrollVertically = bounds.maxTop > 0 && top !== this.uiRef.scrollTop;
 
-			if ((bounds.maxTop > 0 && top !== this.uiRef.scrollTop) || (bounds.maxLeft > 0 && left !== this.uiRef.scrollLeft)) {
+			if (scrollHorizontally || scrollVertically) {
 				this.uiRef.start({
 					targetX: left,
 					targetY: top,
 					animate: (animationDuration > 0) && this.animateOnFocus,
 					duration: animationDuration
 				});
+
+				if (!this.childRef.shouldPreventOverscrollEffect || !this.childRef.shouldPreventOverscrollEffect()) {
+					if (scrollHorizontally) {
+						this.uiRef.updateOverscrollEffect('horizontal', left, overscrollTypes.scrolling, 1);
+					}
+					if (scrollVertically) {
+						this.uiRef.updateOverscrollEffect('vertical', top, overscrollTypes.scrolling, 1);
+					}
+				}
 			}
 			this.lastFocusedItem = item;
 			this.lastScrollPositionOnFocus = pos;
@@ -377,8 +416,14 @@ class ScrollableBase extends Component {
 					this.animateOnFocus = false;
 					Spotlight.focus(nextPage);
 				// Scroll one page with animation if nextPage is equals to `false`
-				} else if (!nextPage) {
+				} else if (nextPage === false) {
 					scrollToAccumulatedTarget(pageDistance, canScrollVertically);
+				} else if (nextPage === null) {
+					const
+						isRtl = this.uiRef.state.rtl,
+						orientation = (direction === 'up' || direction === 'down') ? 'vertical' : 'horizontal',
+						position = (direction === 'up' || !isRtl && direction === 'left' || isRtl && direction === 'right') ? 'before' : 'after';
+					this.uiRef.updateOverscrollEffectByDirection(orientation, position, overscrollTypes.scrolling, 1);
 				}
 			}
 		} else {
@@ -399,8 +444,26 @@ class ScrollableBase extends Component {
 
 	onKeyDown = (ev) => {
 		this.animateOnFocus = true;
-		if (!Spotlight.getPointerMode() && (isPageUp(ev.keyCode) || isPageDown(ev.keyCode)) && !ev.repeat && this.hasFocus()) {
-			this.scrollByPage(ev.keyCode);
+
+		if (!Spotlight.getPointerMode() && !ev.repeat && this.hasFocus()) {
+			if (isPageUp(ev.keyCode) || isPageDown(ev.keyCode)) {
+				this.scrollByPage(ev.keyCode);
+			} else {
+				const direction = getDirection(ev.keyCode);
+				if (direction !== false) {
+					const
+						element = Spotlight.getCurrent(),
+						nextSpottable = element ? getTargetByDirectionFromElement(direction, element) : null;
+
+					if (!nextSpottable) { /* if the spotlight focus will not move */
+						const
+							isRtl = this.uiRef.state.rtl,
+							orientation = (direction === 'up' || direction === 'down') ? 'vertical' : 'horizontal',
+							position = (direction === 'up' || !isRtl && direction === 'left' || isRtl && direction === 'right') ? 'before' : 'after';
+						this.uiRef.updateOverscrollEffectByDirection(orientation, position, overscrollTypes.scrolling, 1);
+					}
+				}
+			}
 		}
 	}
 
@@ -424,6 +487,10 @@ class ScrollableBase extends Component {
 		this.lastFocusedItem = null;
 		this.lastScrollPositionOnFocus = null;
 		this.isWheeling = false;
+		if (this.isVoiceControl) {
+			this.isVoiceControl = false;
+			this.updateFocusAfterVoiceControl();
+		}
 	}
 
 	focusOnItem () {
@@ -485,10 +552,62 @@ class ScrollableBase extends Component {
 		this.uiRef.bounds.scrollHeight = this.uiRef.getScrollBounds().scrollHeight;
 	}
 
+	clearOverscrollEffect = (orientation, position) => {
+		const {type} = this.uiRef.getOverscrollStatus(orientation);
+
+		if (type !== overscrollTypes.none) {
+			this.overscrollJobs[orientation][position].startAfter(
+				(type === overscrollTypes.scrolling) ? overscrollTimeout : 0,
+				orientation,
+				position,
+				overscrollTypes.none,
+				0
+			);
+		}
+	}
+
+	clearAllOverscrollEffects = () => {
+		this.clearOverscrollEffect('horizontal', 'before');
+		this.clearOverscrollEffect('horizontal', 'after');
+		this.clearOverscrollEffect('vertical', 'before');
+		this.clearOverscrollEffect('vertical', 'after');
+	}
+
+	applyOverscrollEffect = (orientation, position, type, ratio) => {
+		const nodeRef = this.overscrollRefs[orientation];
+
+		if (nodeRef) {
+			nodeRef.style.setProperty(overscrollRatioPrefix + orientation + position, ratio);
+
+			if (type === overscrollTypes.scrolling) {
+				this.overscrollJobs[orientation][position].start(orientation, position, overscrollTypes.none, 0);
+			}
+		}
+	}
+
+	newOverscrollJob = (orientation, position) => {
+		if (!this.overscrollJobs[orientation][position]) {
+			this.overscrollJobs[orientation][position] = new Job(this.applyOverscrollEffect.bind(this), overscrollTimeout);
+		}
+	}
+
+	stopOverscrollJob = (orientation, position) => {
+		const job = this.overscrollJobs[orientation][position];
+
+		if (job) {
+			job.stop();
+			this.overscrollJobs[orientation][position] = null;
+		}
+	}
+
 	// FIXME setting event handlers directly to work on the V8 snapshot.
 	addEventListeners = (childContainerRef) => {
 		if (childContainerRef && childContainerRef.addEventListener) {
 			childContainerRef.addEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.addEventListener('webOSVoice', this.onVoice);
+				childContainerRef.setAttribute('data-webos-voice-intent', 'Scroll');
+			}
 		}
 	}
 
@@ -496,6 +615,26 @@ class ScrollableBase extends Component {
 	removeEventListeners = (childContainerRef) => {
 		if (childContainerRef && childContainerRef.removeEventListener) {
 			childContainerRef.removeEventListener('focusin', this.onFocus);
+			if (platform.webos) {
+				childContainerRef.removeEventListener('webOSVoice', this.onVoice);
+				childContainerRef.removeAttribute('data-webos-voice-intent');
+			}
+		}
+	}
+
+	initHorizontalOverscrollRef = (ref) => {
+		if (ref) {
+			this.overscrollRefs.horizontal = ReactDOM.findDOMNode(ref); // eslint-disable-line react/no-find-dom-node
+			this.newOverscrollJob('horizontal', 'before');
+			this.newOverscrollJob('horizontal', 'after');
+		}
+	}
+
+	initVerticalOverscrollRef = (ref) => {
+		if (ref) {
+			this.overscrollRefs.vertical = ref;
+			this.newOverscrollJob('vertical', 'before');
+			this.newOverscrollJob('vertical', 'after');
 		}
 	}
 
@@ -509,6 +648,71 @@ class ScrollableBase extends Component {
 		if (ref) {
 			this.uiRef = ref;
 		}
+	}
+
+	updateFocusAfterVoiceControl = () => {
+		const spotItem = Spotlight.getCurrent();
+		if (spotItem && this.uiRef.containerRef.contains(spotItem)) {
+			const
+				viewportBounds = this.uiRef.containerRef.getBoundingClientRect(),
+				spotItemBounds = spotItem.getBoundingClientRect(),
+				nodes = Spotlight.getSpottableDescendants(this.uiRef.containerRef.dataset.spotlightId),
+				first = this.voiceControlDirection === 'vertical' ? 'top' : 'left',
+				last = this.voiceControlDirection === 'vertical' ? 'bottom' : 'right';
+
+			if (spotItemBounds[last] < viewportBounds[first] || spotItemBounds[first] > viewportBounds[last]) {
+				for (let i = 0; i < nodes.length; i++) {
+					const nodeBounds = nodes[i].getBoundingClientRect();
+					if (nodeBounds[first] > viewportBounds[first] && nodeBounds[last] < viewportBounds[last]) {
+						Spotlight.focus(nodes[i]);
+						break;
+					}
+				}
+			}
+		}
+	}
+
+	onVoice = (e) => {
+		const scroll = e && e.detail && e.detail.scroll;
+		this.isVoiceControl = true;
+
+		switch (scroll) {
+			case 'up':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: true});
+				break;
+			case 'down':
+				this.voiceControlDirection = 'vertical';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: true});
+				break;
+			case 'left':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: true, isVerticalScrollBar: false});
+				break;
+			case 'right':
+				this.voiceControlDirection = 'horizontal';
+				this.onScrollbarButtonClick({isPreviousScrollButton: false, isVerticalScrollBar: false});
+				break;
+			case 'top':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'top'});
+				break;
+			case 'bottom':
+				this.voiceControlDirection = 'vertical';
+				this.uiRef.scrollTo({align: 'bottom'});
+				break;
+			case 'leftmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'left'});
+				break;
+			case 'rightmost':
+				this.voiceControlDirection = 'horizontal';
+				this.uiRef.scrollTo({align: 'right'});
+				break;
+			default:
+				this.isVoiceControl = false;
+		}
+		e.preventDefault();
 	}
 
 	render () {
@@ -535,6 +739,8 @@ class ScrollableBase extends Component {
 			<UiScrollableBase
 				{...rest}
 				addEventListeners={this.addEventListeners}
+				applyOverscrollEffect={this.applyOverscrollEffect}
+				clearAllOverscrollEffects={this.clearAllOverscrollEffects}
 				onFlick={this.onFlick}
 				onKeyDown={this.onKeyDown}
 				onWheel={this.onWheel}
@@ -555,19 +761,19 @@ class ScrollableBase extends Component {
 					rtl,
 					scrollTo,
 					style,
-					touchableProps,
+					touchableProps: {className: touchableClassName, ...restTouchableProps},
 					verticalScrollbarProps
 				}) => (
 					<div
-						className={className}
+						className={classNames(className, overscrollCss.scrollable)}
 						data-spotlight-container={spotlightContainer}
 						data-spotlight-container-disabled={spotlightContainerDisabled}
 						data-spotlight-id={spotlightId}
 						ref={initUiContainerRef}
 						style={style}
 					>
-						<div className={componentCss.container}>
-							<TouchableDiv {...touchableProps}>
+						<div className={classNames(componentCss.container, overscrollCss.overscrollFrame, overscrollCss.vertical, isHorizontalScrollbarVisible ? overscrollCss.horizontalScrollbarVisible : null)} ref={this.initVerticalOverscrollRef}>
+							<TouchableDiv className={classNames(touchableClassName, overscrollCss.overscrollFrame, overscrollCss.horizontal)} ref={this.initHorizontalOverscrollRef} {...restTouchableProps}>
 								{childRenderer({
 									...childComponentProps,
 									cbScrollTo: scrollTo,
@@ -621,14 +827,14 @@ class ScrollableBase extends Component {
  * @ui
  * @public
  */
-const Scrollable = SpotlightContainerDecorator(
+const Scrollable = Skinnable(SpotlightContainerDecorator(
 	{
 		overflow: true,
 		preserveId: true,
 		restrict: 'self-first'
 	},
 	ScrollableBase
-);
+));
 
 export default Scrollable;
 export {