+++ packages/spotlight/SpotlightRootDecorator/SpotlightRootDecorator.js
@@ -99,6 +99,5 @@ const SpotlightRootDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 export default SpotlightRootDecorator;
 export {
-	rootContainerId as spotlightRootContainerName, // DEPRECATED
 	SpotlightRootDecorator
 };