+++ packages/ui/Touchable/Touchable.js
@@ -1,9 +1,9 @@
 /**
- * Provides the [Touchable]{@link ui/Touchable.Touchable} Higher-order Component (HOC) to add
- * gesture support to components and the [configure()]{@link ui/Touchable.configure} method for
- * configuring gestures for the application.
+ * Application support for gestures.
  *
  * @module ui/Touchable
+ * @exports Touchable
+ * @exports configure
  */
 
 import {adaptEvent, call, forward, forwardWithPrevent, forProp, handle, oneOf, preventDefault, returnsTrue} from '@enact/core/handle';
@@ -197,9 +197,9 @@ const defaultConfig = {
 };
 
 /**
- * {@link ui/Touchable.Touchable} is a Higher-order Component that provides a consistent set of
- * pointer events -- `onDown`, `onUp`, and `onTap` -- across mouse and touch interfaces along with
- * support for common gestures including `onFlick`, `onDrag`, onHold`, and `onHoldPulse`.
+ * A Higher-order Component that provides a consistent set of pointer events -- `onDown`, `onUp`,
+ * and `onTap` -- across mouse and touch interfaces along with support for common gestures including
+ * `onFlick`, `onDrag`, `onHold`, and `onHoldPulse`.
  *
  * @class Touchable
  * @memberof ui/Touchable