+++ packages/core/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact core module, newest changes on the top.
 
+## [2.0.0-beta.8] - 2018-06-25
+
+No significant changes.
+
+## [2.0.0-beta.7] - 2018-06-11
+
+No significant changes.
+
 ## [2.0.0-beta.6] - 2018-06-04
 
 No significant changes.